third_party/blink/web_tests/svg/dynamic-updates/SVGMarkerElement-svgdom-markerUnits-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'markerUnits' property of the SVGMarkerElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS markerElement.markerUnits.baseVal is SVGMarkerElement.SVG_MARKERUNITS_USERSPACEONUSE
PASS markerElement.markerUnits.baseVal is SVGMarkerElement.SVG_MARKERUNITS_STROKEWIDTH
PASS successfullyParsed is true

TEST COMPLETE
